backend/requirements.txt
alembic==1.12.1
decorator==5.1.1
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.1
pandas==2.1.3
pbr==6.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
Tempita==0.5.2
typing_extensions==4.8.0
tzdata==2023.3
